+++ lib/src/test.dart
@@ -1,7 +1,0 @@
-/*
- * Copyright (c) 2019-2020 unTill Pro, Ltd. and Contributors
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-