+++ CHANGELOG.md
@@ -3,3 +3,12 @@
 - "End of source" error type handling added
 - Options now can be passed to a Tagser constructor.
 - `ignoreCase` option available now. If `true` tag name spelling will be ignored.
+
+## 1.0.2
+
+- Empty valued attributes now available. Example: ` attr="" `
+- new Tagser method - `.html(tags)` - where `tags` is a list of tags from `.parse()` method. This method returns correct displayable HTML for any tags;
+
+## 1.0.3
+
+- Now with null-safety